move InlinerOfUserOper and LetInExpander to tla-pp
