+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/FoldSeqRule.scala
@@ -6,8 +6,8 @@ import at.forsyte.apalache.tla.lir.convenience.tla
 import at.forsyte.apalache.tla.lir.oper.ApalacheOper
 import at.forsyte.apalache.tla.lir.storage.BodyMapFactory
 import at.forsyte.apalache.tla.lir.transformations.impl.IdleTracker
-import at.forsyte.apalache.tla.lir.transformations.standard.InlinerOfUserOper
 import at.forsyte.apalache.tla.lir._
+import at.forsyte.apalache.tla.pp.InlinerOfUserOper
 
 /**
  * Rewriting rule for FoldSeq.
