+++ tla-pp/src/test/scala/at/forsyte/apalache/tla/pp/TestLetInExpander.scala
@@ -1,11 +1,11 @@
-package at.forsyte.apalache.tla.lir
+package at.forsyte.apalache.tla.pp
 
 import at.forsyte.apalache.tla.lir.UntypedPredefs._
 import at.forsyte.apalache.tla.lir.convenience.tla
 import at.forsyte.apalache.tla.lir.oper.{TlaOper, TlaSeqOper}
 import at.forsyte.apalache.tla.lir.transformations.impl.IdleTracker
-import at.forsyte.apalache.tla.lir.transformations.standard._
 import at.forsyte.apalache.tla.lir.values.TlaInt
+import at.forsyte.apalache.tla.lir._
 import org.junit.runner.RunWith
 import org.scalatest.FunSuite
 import org.scalatest.junit.JUnitRunner
